pyproject.toml
[project]
name = "monitor"
version = "0.1.0"
description = "測試監控系統 - PHP 重構版"
requires-python = ">=3.11"
dependencies = [
    # Web 框架
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # 認證與安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    # DB (僅認證 DB 用)
    "sqlalchemy>=2.0.35",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    # Celery
    "celery>=5.4.0",
    "redis>=5.1.0",
    # 監控
    "prometheus-client>=0.21.0",
    # 工具
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
